[기능추가] 심볼릭링크 함수를 FileControl으로 이동 및 기능 작동되게 구현 | 메타데이터 출력 기능 미작동 개선
